# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/comments/{id}": {
      "patch": {
        "operationId": "CommentController_updateComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfull."
          },
          "400": {
            "description": "Bad Request."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Comment not found."
          }
        },
        "summary": "Update comment.",
        "tags": [
          "Comment"
        ]
      },
      "delete": {
        "operationId": "CommentController_deleteComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successfull."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Comment not found."
          }
        },
        "summary": "Delete comment.",
        "tags": [
          "Comment"
        ]
      }
    },
...
    "/notifications": {
      "get": {
        "operationId": "NotificationController_getNotification",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "isRead",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfull."
          }
        },
        "summary": "Get user notification.",
        "tags": [
          "Notification"
        ]
      }
    },
    "/notifications/{id}/read": {
      "patch": {
        "operationId": "NotificationController_ReadSingleNotification",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfull."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Comment not found."
          }
        },
        "summary": "Mark a notification as read.",
        "tags": [
          "Notification"
        ]
      }
    },
...
  },
  "info": {
    "title": "Blog API",
    "description": "API Documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateCommentDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}